+++ src/Runtime/Optimized/wasm_src/src/mmd_morph_controller.rs
@@ -32,7 +32,7 @@ impl MmdMorphController {
                     };
                     let index = morph_indices[i];
 
-                    if group_morph_stack.iter().any(|(index, _)| *index == morph_index) {
+                    if group_morph_stack.iter().any(|(x, _)| *x == index) {
                         // no need to diagnostic message because same error will be reported in the typescript side
                         morph_indices[i] = -1;
                     } else if 0 <= index && index < morphs.len() as i32 {